extracting accounts dropdown menu component
