+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/operations/components/AccountsDropdownMenu.kt
@@ -1,0 +1,171 @@
+/*
+ *
+ * Copyright (C) 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Modifications made by Roberto Kenzo Hamano, 2024
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package br.com.rbrthmn.ui.financialcompanion.screens.operations.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowDropDown
+import androidx.compose.material.icons.filled.Check
+import androidx.compose.material.icons.filled.Close
+import androidx.compose.material3.Button
+import androidx.compose.material3.Card
+import androidx.compose.material3.DropdownMenu
+import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.window.Dialog
+import br.com.rbrthmn.R
+
+sealed class AccountType(val stringId: Int)
+data object AimedAccount : AccountType(R.string.aimed_account_hint)
+data object OriginAccount : AccountType(R.string.origin_account_hint)
+
+@Composable
+fun AccountsDropdownMenu(modifier: Modifier = Modifier, accountType: AccountType) {
+    var expanded by remember { mutableStateOf(false) }
+    var showAddAccountDialog by remember { mutableStateOf(false) }
+    val options = listOf(
+        "Conta A",
+        "Conta B",
+        "Conta C",
+        "Conta D",
+        "Conta E",
+        "Conta F",
+        "Conta G",
+    ).plus(stringResource(id = R.string.add_account_option))
+    var selectedOptionText: String? by remember { mutableStateOf(null) }
+
+    if (showAddAccountDialog) {
+        AddSimpleAccountDialog(
+            onCancelButtonClick = { showAddAccountDialog = false },
+            onSaveButtonClick = { showAddAccountDialog = false })
+    }
+
+    Column(modifier = modifier) {
+        OutlinedTextField(
+            value = selectedOptionText ?: stringResource(id = R.string.blank),
+            label = { Text(text = stringResource(id = accountType.stringId)) },
+            onValueChange = { selectedOptionText = it },
+            readOnly = true,
+            trailingIcon = {
+                IconButton(onClick = { expanded = true }) {
+                    Icon(Icons.Filled.ArrowDropDown, "contentDescription")
+                }
+            },
+            singleLine = true
+        )
+        DropdownMenu(
+            expanded = expanded,
+            onDismissRequest = { expanded = false }
+        ) {
+            options.forEach { selectedOption ->
+                DropdownMenuItem(
+                    onClick = {
+                        selectedOptionText = selectedOption
+                        expanded = false
+                        if (selectedOptionText == options.last()) showAddAccountDialog = true
+                    },
+                    text = { Text(text = selectedOption) },
+                    leadingIcon = {
+                        Icon(
+                            painter = painterResource(id = R.drawable.help),
+                            contentDescription = "Balance Icon"
+                        )
+                    }
+                )
+            }
+        }
+    }
+}
+
+@Composable
+private fun AddSimpleAccountDialog(
+    modifier: Modifier = Modifier,
+    onCancelButtonClick: () -> Unit,
+    onSaveButtonClick: () -> Unit
+) {
+    var accountHolder by remember { mutableStateOf("") }
+
+    Dialog(onDismissRequest = onCancelButtonClick) {
+        Card(
+            modifier = modifier.fillMaxWidth()
+        ) {
+            Column(
+                horizontalAlignment = Alignment.CenterHorizontally,
+                modifier = modifier.padding(
+                    horizontal = dimensionResource(id = R.dimen.padding_large),
+                    vertical = dimensionResource(id = R.dimen.padding_medium)
+                )
+            ) {
+                OutlinedTextField(
+                    label = { Text(text = stringResource(id = R.string.account_holder_hint)) },
+                    value = accountHolder,
+                    onValueChange = { accountHolder = it },
+                )
+                Row(
+                    modifier = modifier
+                        .fillMaxWidth()
+                        .padding(top = dimensionResource(id = R.dimen.padding_small)),
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.SpaceAround,
+                ) {
+                    Button(onClick = onCancelButtonClick) {
+                        Icon(
+                            imageVector = Icons.Default.Close, contentDescription = stringResource(
+                                id = R.string.close_icon_description
+                            )
+                        )
+                    }
+                    Button(onClick = onSaveButtonClick) {
+                        Icon(
+                            imageVector = Icons.Default.Check, contentDescription = stringResource(
+                                id = R.string.check_icon_description
+                            )
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
+
+@Preview
+@Composable
+fun AddSimpleAccountDialogPreview(modifier: Modifier = Modifier) {
+    AddSimpleAccountDialog(onSaveButtonClick = {}, onCancelButtonClick = {})
+}
